refactor: rename LAN service
