+++ services/cloudapi/cloudapilan/lan.go
@@ -1,4 +1,4 @@
-package lanSvc
+package cloudapilan
 
 import (
 	"context"
